+++ packages/git-mob-core/src/git-mob-api/git-config.ts
@@ -1,4 +1,4 @@
-import { getConfig } from './exec-command.js';
+import { getConfig, setConfig } from './exec-command.js';
 
 export async function getLocalCommitTemplate() {
   return getConfig('--local commit.template');
@@ -14,4 +14,12 @@ export async function getGitUserName() {
 
 export async function getGitUserEmail() {
   return getConfig('user.email');
+}
+
+export async function setGitUserName(name: string) {
+  return setConfig('user.name', name);
+}
+
+export async function setGitUserEmail(email: string) {
+  return setConfig('user.email', email);
 }